+++ src/Controller/IuForm/Utils/StartData.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Controller\IuForm\IuFormUtils;
+namespace App\Controller\IuForm\Utils;
 
 // No validation here. If you play with the scripts and something goes wrong, blame yourself.
 class StartData
